requirements.txt
python-dotenv
requests
beautifulsoup4
langchain
langchain-openai
langgraph
